ticket-4.2/living_room_dicover update
