+++ lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/foundation.dart' show kIsWeb;
+import 'package:saloonprived_app/src/screens/adult_warning/adult_warning_screen.dart';
 import 'package:url_strategy/url_strategy.dart';
 
 import 'src/app.dart';
@@ -22,7 +23,7 @@ void main() async {
       useFallbackTranslations: true,
       supportedLocales: supportedLocales,
       fallbackLocale: supportedLocales.first,
-      child: const SaloonprivedApp(),
+      child: const AttentionPageScreen(),
     ),
   );
 }
